# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: EmotionSNS API
  version: 1.0.0
servers:
  - url: https://api.emotionsns.example.com
paths:
  /posts:
    post:
      summary: Create a new post with emotion analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              example:
                postId: "abc123"
                status: "created"
        "400":
          description: Invalid request
          content:
            application/json:
              example:
                error: "Missing userId or content"
...
  /users/{userId}/feed:
    get:
      summary: Get user's post feed, optionally filtered by emotion
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: emotion
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of feed posts
          content:
            application/json:
              example:
                posts:
                  - postId: "abc123"
                    userId: "userX"
                    content: "今日は最高！"
                    emotionTags:
                      - type: "joy"
                        score: 0.9
...